requirements.txt
altair==4.2.0
numpy==1.24.1
pandas==1.5.2
protobuf==4.22.0
streamlit==1.18.1
db-dtypes==1.0.5
google-auth==2.16.1
google-cloud-bigquery==3.5.0